# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: X-Ray Academy AI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /research/generate:
    post:
      operationId: generateResearch
      tags: [research]
      summary: Stream AI-generated research content (SSE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchInput"
      responses:
        "200":
          description: Server-Sent Event stream of generated content
          content:
            text/event-stream:
              schema:
                type: string
...
  /research/outputs/{outputId}:
    delete:
      operationId: deleteResearchOutput
      tags: [research]
      summary: Delete a saved research output
      parameters:
        - name: outputId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ResearchInput:
      type: object
      required: [mode, topic]
      properties:
        mode:
          type: string
          enum:
            - paper_ieee
            - paper_elsevier
            - literature_review
            - research_ideas
            - research_gaps
            - experiment_plan
            - patent_analysis
            - technical_report
            - security_algorithm
        topic:
          type: string
        context:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        word_count:
          type: ["integer", "null"]
...